pyproject.toml
[tool.poetry]
name = "qq-bot-openclaw"
version = "1.0.0"
description = "QQ Bot with OpenClaw AI Assistant"
authors = ["OpenClaw"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.0.0"
httpx = "^0.24.0"
aiohttp = "^3.8.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
pydantic = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.0.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["plugins"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
